requirements-dev.txt
pytest==8.4.2
PyYAML==6.0.2
ruff==0.13.0
